hrv: widen pct's multiply so multiMs stops reporting a negative percentage
`deliveryHistogram` calls `pct(msToInt(multiMs), msToInt(knownMs))`, where both
arguments are beat-time in milliseconds summed over a whole night rather than a
row count. `part * 200` was evaluated in a 32-bit `Int` and overflows above
2147483647 / 200 = 10,737,418 ms, i.e. 2.98 h of beat-time on multi-delivery
seconds; a real over-count night carries three to four times that.

Kotlin's `Int` wraps silently, so `multiMs` was reported negative on every such
night. It is a share of a positive subset of a positive total and cannot be
negative. Swift's `Int` is 64-bit on the platforms that run it today, so the
same source computed the right answer there and the two twins disagreed.

Widen the operation rather than the signature, so no caller changes and the
integer half-up tie behaviour is untouched. The Swift twin states `Int64`
explicitly rather than inheriting the platform width, since StrandAnalytics also
builds for watchOS, where `Int` is 32-bit.

The same guard already exists thirty lines below in `duplicatePairRatios` and
again in `V18AuxCodec`; `pct` predates both.

Adds a regression case at night scale to the file that already pins `pct`. The
existing tie cases are unchanged by the widening.

+++ Packages/StrandAnalytics/Sources/StrandAnalytics/HRVAnalyzer.swift
@@ -808,8 +808,15 @@ public enum HRVAnalyzer {
     static func msToInt(_ ms: Double) -> Int { ms > 0 ? Int(ms + 0.5) : 0 }
 
     /// Whole-percent, integer half-up, so both platforms round a tie the same way. 0 when `total` is 0.
+    ///
+    /// The arithmetic is widened to `Int64` because `deliveryHistogram` passes BEAT-TIME IN
+    /// MILLISECONDS summed over a whole night, not a row count. `part * 200` needs more than 32 bits
+    /// above 10,737,418 ms (2.98 h of beat-time on multi-delivery seconds) and a real over-count night
+    /// carries 3-4x that. `Int` is 64-bit on iOS/macOS so this platform read correctly while the Kotlin
+    /// twin wrapped to a negative percentage; it is 32-bit on `arm64_32`, and StrandAnalytics already
+    /// builds for watchOS, so the width is stated rather than inherited. Twin of Kotlin `pct`.
     static func pct(_ part: Int, _ total: Int) -> Int {
-        total > 0 ? (part * 200 + total) / (total * 2) : 0
+        total > 0 ? Int((Int64(part) * 200 + Int64(total)) / (Int64(total) * 2)) : 0
     }
 
     /// #1008: a compact, deterministic RAW-ROW sample of the beats around the DENSEST second, for the
